contracts.py
from solcx import compile_standard, install_solc
import json
import os

# Install Solidity compiler version 0.8.0
install_solc('0.8.0')

# Read the Solidity contract
with open('contracts/MyContract.sol', 'r') as file:
    source_code = file.read()

# Compile the contract
compiled_sol = compile_standard(
    {
        "language": "Solidity",
        "sources": {
            "MyContract.sol": {"content": source_code}
        },
        "settings": {
            "outputSelection": {
                "*": {
                    "*": ["abi", "metadata", "evm.bytecode", "evm.sourceMap"]
                }
            }
        },
    },
    solc_version="0.8.0",
)

# Create 'build' directory if not exists
if not os.path.exists('build'):
    os.makedirs('build')

# Save compiled output
with open('build/compiled_contract.json', 'w') as f:
    json.dump(compiled_sol, f, indent=4)

# Extract ABI and Bytecode

abi = compiled_sol['contracts']['MyContract.sol']['CyberInsurance']['abi']
bytecode = compiled_sol['contracts']['MyContract.sol']['CyberInsurance']['evm']['bytecode']['object']

print("Contract ABI and bytecode compiled successfully!")

